+++ src/components/addPhoto/endingMap/EndingMap.tsx
