+++ src/components/ending/carousel/Carousel.tsx
