+++ src/components/ending/totalPay/TotalPay.tsx
