+++ src/components/plan/updatePlan/MapModal.tsx
@@ -4,15 +4,14 @@ import { type SubmitHandler, useForm } from 'react-hook-form';
 import { useParams } from 'react-router-dom';
 
 import { type PinContentsType, addPin, updatePin } from '@api/pins';
+import MapModalInput from '@components/plan/addPlan/MapModalInput';
+import MapNonePoly from '@components/plan/addPlan/MapNonePoly';
+import MapModalLayout from '@components/plan/addPlan/ModalLayout';
 import useConfirm from '@hooks/useConfirm';
 import { updatePinStore } from '@store/updatePinStore';
 import { uuid } from '@supabase/gotrue-js/dist/module/lib/helpers';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import _ from 'lodash';
-
-import MapModalInput from '../MapModalInput';
-import MapNonePoly from '../MapNonePoly';
-import MapModalLayout from '../ModalLayout';
 
 interface InputType {
   address?: string;
