+++ src/pages/NotFound.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import NotFoundFull from '@components/404/NotFoundFull';
+import NotFoundFull from '@components/notfound/NotFoundFull';
 
 const NotFound: React.FC = () => {
   return <NotFoundFull />;
